services/accounts/requirements.txt
-i https://pypi.org/simple
google-api-core==1.3.0
grpcio==1.14.2
mysql-connector==2.2.9
protobuf==3.6.1
pymongo==3.9.0
